John - Implemented EnemyVisibility
Added an EnemyVisibility script that allows enemies with the script attached to be able to draw a cone in-front of themselves so that they can see the player if they are 1) in front of them in a specific distance and 2) within their cone. May merge with the GuardAI script so that guards can see the player and react according to different logic (whether they are peaceful, alert, attacking, etc).