Guard AI + Player Hide = Completed
Fixed the guardAi, enemyVisibility, and followguard scripts. Player can now hide and the guard's visibilty turns green and stops moving, if the player is close then it turns yellow still not moving, but if the player is too close then the guard chases the player and it  turns red. Capsule is above the guard head and changes colors automatically.