John - Added Crouching Mechanic
Crouching now scales the player model down to 0.65 their size, allowing them to evade guard vision lines when crouched (probably will change either the scale to 0.75 or the guard vision cones to also be able to adjust for height changes made by the player).

Note: Currently the player floats a bit above the ground instead of fully falling to the floor. It doesn't appear to cause any issues, but we can resolve this later.