John - Adjusted Rocks + Adjusted Player Height
Rocks can now be destroyed when they collide with something (a wall or an object with collisions). The player scaling was changed from 0.65 to 0.75. May need more adjusting to allow guards to still see them even if they're crouching right in-front of them (guards can't see the player when the player suddenly crouches in front of them).