John - Fixed issue with PauseMenu
Issue that was fixed: When the pause menu (ESC key) was pressed and the player was out of dialogue, the player could still look around using their mouse and instantiate rocks using left click (Mouse 0).

Files changed: PlayerController.cs (added a reference to the pause menu, then changed the condition for canMove being set to true to also include whether or not the pause menu was enabled). ThrowProjectiles (same deal, added a reference to the pause menu, then added an additional check to see if the pause menu was open so that if it was, rocks could not be thrown).