John - Added SecondFloorTutorial objects + script
Major Issues: Currently, it appears as though the SecondFloorTutorialManager object and the SecondFloorTutorial script are not function *period*. As in, when you press play, although the objects are present in the scene (even the invisible colliders + triggers, and the visible triggers for debugging), nothing seems to be working.

What should be occurring: When the player gets to the second floor, they interact with a trigger that enables a Box Collider behind them. The Box Collider prevents the player from leaving the second floor until all four of the separate Trigger cubes have been activated in the four rooms on the second floor. Once the player activates all four trigger cubes, the box collider at the stairwell disables, allowing the player to leave.

For testing: Could someone try looking into why the SecondFloorTutorial script isn't currently working? I'm really confused why it isn't.