John - Fixed Second Floor Tutorial
Currently works as it should. After the player interacts with the trigger zone upstairs, the box collider is enabled behind them. Should the player interact with all four trigger cubes within each of the four rooms, the box collider will be disabled, allowing the player to exit the second floor.

Changes to be made: Instead of trigger zone cubes, we'll need to add the other elements of the tutorial (press E to interact with something, throw rocks, speak with NPCs, escape the tavern).