John - Added Ability to Steal Stuff!
We now have the ability to steal stuff. So long as an object has the correct tag "StealableItem" within the tag dropdown of the inspector, the player can now press E when nearby one to collect that item, increasing a gold counter.

Things to add: A UI element showing the current gold count. Right now it only sends it to the Console for debugging purposes.