John - Added/Tweaked NPC Animations
For any new NPC or any existing ones, so long as we attach the NPCAnimator script as well as the NPCAnimator controller, the NPCs should properly decide when to play animations based on whether or not they're the activeNPC being spoken to. If they are, every time the continue button is pressed, they'll play the talking animation.

Note: We'll probably need to tweak the NPCAnimator, DialogManager, and PlayerController scripts later to ensure that if the player activates a new dialogue with some NPC, then that NPC becomes the new activeNPC, and then the new NPC can play the speaking animation.