John - Added Rock_Bag item
Allows the player to pick up a Rock Bag that gives the player 5 rocks. To test, instantiate a Rock_Bag from the Prefab into the scene. Walk into it with the player, Press E, and the 5 rocks will be added to your total amount of rocks.

Also adjusted the radius of the stealable items in Prefabs, so they're a bit closer to the item. May adjust later to be larger/smaller depending on where we want items to be. May also had a new Prefab folder for "NonStealableItems" so that the player can't just steal *anything*.