John - Outside Escape Tutorial Finished
Added all of the different obstacles, guards, rock bags, and lootable items to the outside!

Current issues: Guards can run through walls if a rock is used to bait them. There also just isn't *enough* in the area, but that's okay. We'll work on changing this area later, if we want to.