John - Fixed guards being able to phase through objects (NEEDS MORE WORK)
Fix: I adjusted the GuardAI script to ensure that they should not be able to phase through walls and objects if they can't navigate around/through them. The only issue is that we need to add NavMesh Obstacle components to each of the walls and objects (should be easily done through the prefabs for that asset package) and re-bake the NavMesh.

Treasure, you'll have to re-do the NavMesh. I can handle adding the NavMesh Obstacle components to the walls + objects after that's re-baked.