John - Working on Adding Flexible Dialogue System
Status: Active ... Currently on video #5 of Semag Games tutorial. May continue, but his video focuses on a 2D environment as opposed to 3D interactions.

Will need to add 3D interactions between the player and the character... or have the campfire stories be a 2D scene between the player and the NPC.