Hiking Scene is working!
Health system is good to go!
Interaction with animals is good to go!
Log + Damage system is good to go!
Win/Lose conditions are good to go!
Colliders are good to go!

Note: I need to work on animation, music/sound, and tutorial text for the beginning of the scene.