Added a platform script
I removed this script because I thought it would be better for the text to be on the dialogue panel instead of above the platform's location. This will give the player better visually and let the response buttons control if the player can go to the mini game or stay.